@@ -1,12 +1,13 @@
 {
   "name": "angular-iscroll-probe",
-  "version": "4.0.0",
+  "version": "4.0.2",
   "description": "A version of angular-iscroll that depends on iscroll-probe.js instead of iscroll.js.",
   "main": "dist/lib/angular-iscroll.js",
   "scripts": {
     "build": "NODE_ENV=production webpack --config-name=library",
     "build-dev": "NODE_ENV=development webpack --config-name=library",
     "watch": "NODE_ENV=development webpack --config-name=library --watch",
+    "update-browserlist-db": "npx update-browserslist-db@latest",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
@@ -35,29 +36,34 @@
   "peerDependencies": {
     "angular": ">=1.8"
   },
+  "resolutions": {
+    "loader-utils": "^2.0.4",
+    "minimatch": "^3.0.8"
+  },
   "devDependencies": {
-    "@babel/core": "^7.14.6",
-    "@babel/preset-env": "^7.14.7",
-    "@uirouter/angularjs": ">=1.0.29",
-    "babel-loader": "^8.2.2",
+    "@babel/core": "^7.23.5",
+    "@babel/preset-env": "^7.23.5",
+    "@uirouter/angularjs": "^1.1.0",
+    "babel-loader": "^9.1.3",
     "babel-plugin-angularjs-annotate": "^0.10.0",
-    "bootstrap-sass": "^3.4.1",
-    "clean-webpack-plugin": "^3.0.0",
-    "compression-webpack-plugin": "^8.0.1",
-    "css-loader": "^5.2.6",
-    "cssnano": "^5.0.6",
-    "dateformat": "^4.5.1",
+    "bootstrap-sass": "^3.4.3",
+    "clean-webpack-plugin": "^4.0.0",
+    "compression-webpack-plugin": "^10.0.0",
+    "css-loader": "^6.8.1",
+    "cssnano": "^6.0.1",
+    "dateformat": "^5.0.3",
     "file-loader": "^6.2.0",
-    "html-loader": "^2.1.2",
-    "jquery": "^3.6.0",
-    "jshint": "^2.13.0",
-    "mini-css-extract-plugin": "^1.6.1",
-    "node-sass": "^6.0.1",
-    "postcss-loader": "^6.1.0",
-    "postcss-preset-env": "^6.7.0",
-    "sass-loader": "^12.1.0",
+    "html-loader": "^4.2.0",
+    "jquery": "^3.7.1",
+    "jshint": "^2.13.6",
+    "mini-css-extract-plugin": "^2.7.6",
+    "postcss-loader": "^7.3.3",
+    "postcss-preset-env": "^9.3.0",
+    "sass": "^1.69.5",
+    "sass-loader": "^13.3.2",
     "url-loader": "^4.1.1",
-    "webpack": "^5.40.0",
-    "webpack-cli": "^4.7.2"
-  }
+    "webpack": "^5.89.0",
+    "webpack-cli": "^5.1.4"
+  },
+  "packageManager": "yarn@4.0.2"
 }
